+++ src/com/yanncebron/m68kplugin/lang/psi/M68kTokenGroups.java
@@ -236,6 +236,7 @@ public final class M68kTokenGroups {
     OFFSET,
     OPT,
     ORG,
+    OUTPUT,
     PAGE,
     PLEN,
     POPSECTION,