+++ src/com/yanncebron/m68kplugin/lang/psi/M68kTokenTypes.java
@@ -242,6 +242,7 @@ public abstract class M68kTokenTypes {
   public static final IElementType OR = new M68kTokenType("or");
   public static final IElementType ORG = new M68kTokenType("org");
   public static final IElementType ORI = new M68kTokenType("ori");
+  public static final IElementType OUTPUT = new M68kTokenType("output");
   public static final IElementType PAGE = new M68kTokenType("page");
   public static final IElementType PC = new M68kTokenType("pc");
   public static final IElementType PEA = new M68kTokenType("pea");