+++ tests/com/yanncebron/m68kplugin/parser/DirectivesParsingTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 The Authors
+ * Copyright 2022 The Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -579,6 +579,10 @@ public class DirectivesParsingTest extends M68kParsingTestCase {
 
   public void testMask2Directive() throws Exception {
     doCodeTest(" mask2");
+  }
+
+  public void testOutputDirective() throws Exception {
+    doCodeTest(" output fileName");
   }
 
 }